chore: simplify and move ide code
