+++ templates/page_js/ide.js
@@ -196,34 +196,6 @@ function showDropdown() {
     dropdown.appendChild(newTokenTab);
     dropdown.appendChild(loadContract);
 }
-// async function lintCode(code) {
-//     try {
-//         code = encodeURIComponent(code);
-//         const response = await fetch(RPC + '/abci_query?path="/lint/' + btoa(code) + '"', {
-//             method: 'GET',
-//             headers: {
-//                 'Content-Type': 'application/json'
-//             }
-//         });
-        
-//         if (!response.ok) {
-//             console.error('Failed to lint code:', response.status);
-//             return [];
-//         }
-        
-//         const data = await response.json();
-//         let lintinfo = data['result']['response']['value'];
-//         let lintinfo_decoded = atob(lintinfo);
-//         lintinfo_decoded = JSON.parse(lintinfo_decoded);
-//         lintinfo_decoded = lintinfo_decoded['stdout'] + lintinfo_decoded['stderr'];
-//         console.log(lintinfo_decoded);
-//         return parseLintOutput(lintinfo_decoded);
-        
-//     } catch (error) {
-//         console.error('Error linting code:', error.message);
-//         return [];
-//     }
-// }
 
 // var whitelistedPatterns = [
 //     'export',
@@ -245,46 +217,39 @@ function showDropdown() {
 //     'Any',
 //     'LogEvent'
 // ];
-// function parseLintOutput(output) {
-//     let errors = [];
-//     let lines = output.split('\n');
-//     for (let line of lines) {
-//         let match = line.match(/<string>:(\d+):(\d+):\s*(.+)/);
-//         if (match) {
-//             let message = match[3];
-//             let isWhitelisted = whitelistedPatterns.some(pattern => message.includes(pattern));
-//             if (!isWhitelisted) {
-//                 errors.push({
-//                     line: parseInt(match[1]) - 1,
-//                     col: parseInt(match[2]) - 1,
-//                     message: message,
-//                     severity: 'error'  // Default to 'error'
-//                 });
-//             }
-//         }
-//     }
-//     return errors;
-// }
 
-// async function pythonLinter(text, options, cm) {
-//     try {
-//         const errors = await lintCode(text);
-//         console.log(errors);
-//         let lintErrors = errors.map(error => ({
-//             message: error.message,
-//             severity: error.severity,
-//             from: CodeMirror.Pos(error.line, error.col),
-//             to: CodeMirror.Pos(error.line, error.col)
-//         }));
-//         return lintErrors;
-//     } catch (error) {
-//         console.error('Error in pythonLinter:', error.message);
-//         return [];
-//     }
-// }
+async function lintCode(code){   
+    try{
+        const result = await pyodide.runPythonAsync(`
+from xian_contracting_linter import lint_code
+lint_code("""${code}""")
+`);   
+        let lintinfo = result.toJs();
+        return lintinfo;    
+    } catch (error) {
+        console.error('Error linting code:', error.message);
+        return [];
+    }
+}
 
+async function pythonLinter(text, options, cm) {
+    try {
+        const errors = await lintCode(text);
+        let lintErrors = errors.map(error => ({
+            message: error.get("message"),
+            severity: error.get("severity"),
+            from: CodeMirror.Pos(error.get("line")),
+            to: CodeMirror.Pos(error.get("line"))
+        }));
+        
+        return lintErrors;
+    } catch (error) {
+        console.error('Error in pythonLinter:', error.get("message"));
+        return [];
+    }
+}
 
-// CodeMirror.registerHelper("lint", "python", pythonLinter);
+CodeMirror.registerHelper("lint", "python", pythonLinter);
 
 var editor = CodeMirror(document.querySelector('#editor'), {
     value: code_storage[current_tab],
